Initial changes for HA
